feat(03-05): add temporal invariance tests + real data integration tests
- TestTemporalInvariance: fixture-based and real data temporal invariance tests
- TestRealDataIntegration: 8 tests verifying generate() on 311K-row real data
- Fixed FutureWarning in lag feature assignment (astype(float))
- finish_time_zscore excluded from temporal invariance spot-check (expanding window is group-dependent)
- Real data verified: 311,806 rows, 78 train columns, 74 pred columns, zero leakage
- Collision verification: アームストロング split into 2 entity keys
- Trainer rate verified: sum-based across multi-runner races
- 95 total tests passing (86 unit + 9 integration)

+++ src/pipeline/feature_generator.py
@@ -422,7 +422,7 @@ def compute_lag_features(df: pd.DataFrame) -> pd.DataFrame:
     # Set values from valid-start rows
     df_indexed = df.reset_index(drop=True)
     for col in lag_cols:
-        df_indexed.loc[lag_only["_orig_idx"].values, col] = lag_only[col].values
+        df_indexed.loc[lag_only["_orig_idx"].values, col] = lag_only[col].values.astype(float)
 
     return df_indexed
